Fix deletion worker communication protocol
The sai_deletion_worker was failing to delete old job directories
because the communication protocol over the pipe was not robust.
Directory names were written to the pipe without delimiters, causing the
worker to sometimes read multiple names concatenated together, leading
to invalid paths.

This commit fixes the protocol by:
1. Appending a newline character to each directory name written to the
   pipe in `scan_jobs_dir_cb`.
2. Implementing line-buffered reading in `sai_deletion_worker` to
   correctly parse the newline-delimited stream of directory names.

This ensures that the deletion worker processes one directory at a time,
even when deletion requests are sent in quick succession.

+++ src/builder/b-task.c
@@ -412,7 +412,7 @@ saib_task_destroy(struct sai_nspawn *ns)
 		{
 			DWORD written;
 			if (!WriteFile(builder.pipe_master_wr_win, ns->task->uuid,
-				       strlen(ns->task->uuid), &written, NULL))
+				       (DWORD)strlen(ns->task->uuid), &written, NULL))
 				lwsl_err("%s: failed to write to deletion worker\n",
 					 __func__);
 		}
